# openapi.yaml
openapi: 3.0.3
info:
  title: University Clearance System API
  version: 1.0.0
  description: API contract for department, student, clearance, approval, notification, and authentication endpoints.
servers:
  - url: http://localhost:3000/api
paths:
  /departments:
    get:
      summary: Get all departments
      responses:
        '200':
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
    post:
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        '201':
          description: Created department
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
  /departments/{id}:
    get:
      summary: Get a department by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
    put:
      summary: Update a department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        '200':
          description: Updated department
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
    delete:
      summary: Delete a department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Department deleted
  /students:
    get:
      summary: Get all students
      responses:
        '200':
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '201':
          description: Created student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
  /students/{id}:
    get:
      summary: Get a student by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    put:
      summary: Update a student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Updated student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    delete:
      summary: Delete a student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Student deleted
  /clearance-requests:
    get:
      summary: Get all clearance requests
      responses:
        '200':
          description: List of clearance requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClearanceRequest'
    post:
      summary: Create a clearance request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClearanceRequest'
      responses:
        '201':
          description: Created clearance request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClearanceRequest'
...
components:
  schemas:
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contact:
          type: string
        staff:
          type: array
          items:
            type: string
        approvalSettings:
          type: object
    Student:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        departmentId:
          type: string
        clearanceStatus:
          type: string
        documents:
          type: array
          items:
            type: string
    ClearanceRequest:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        departmentId:
          type: string
        status:
          type: string
        history:
          type: array
          items:
            type: string
...